selenium/requirements.txt
pillow
requests
face_recognition
dlib
cmake
block_distortion
bs4
tqdm